+++ packages/smooth-dnd/src/interfaces.ts
@@ -1,4 +1,4 @@
-import { ContainerOptions } from './exportTypes';
+import { ContainerOptions, DropIndicator } from './exportTypes';
 
 export type Position = { x: number; y: number };
 export type MousePosition = { clientX: number; clientY: number };
@@ -45,6 +45,7 @@ export interface DragResult {
   shadowBeginEnd: Boundary & { beginAdjustment: number; rect?: TopLeft, dropArea?: Boundary };
   containerBoxChanged?: boolean;
   dropPlaceholderContainer?: HTMLDivElement;
+  dropIndicator?: DropIndicator | null;
 }
 
 export interface DraggableInfo {